requirements.txt
boto3==1.38.3
botocore==1.38.3
fastapi==0.115.12
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.3
PyJWT==2.10.1
python-dotenv==1.1.0
Requests==2.32.3
uvicorn==0.34.2